requirements.txt
psycopg[binary]>=3.3
python-dotenv>=1.0
playwright>=1.40
warcio>=1.7
opentimestamps-client>=0.7
requests>=2.31
boto3>=1.34
